internal: add involved object api groups filter
